+++ ui_to_py.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+files=`ls homura_art/ui/*.ui`
+for file in $files
+do
+   pyside6-uic "${file%%.*}".ui -g python -o "${file%%.*}".py
+done